Firebase Analytics, Crashytics, Messaging 기능 추가

+++ BicycleRecord/Resources/Constants/Weather.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct Weather {
     static var wea1: (String, String, Double)?
-    static var wea2: Int?
+    static var wea2: Double?
     static var wea3: (Double, Double)?
     static var dong: String?
 }
